Added routes to express server
