+++ @types/Param.ts
@@ -1,0 +1,4 @@
+export default interface Param {
+    a: number,
+    b: number
+}
